[security] RTT: Real-Time Targeting
To achieve faster response to hashrate changes and improve stability of
inter-block times, a new and radically different difficulty adjustment
algorithm is introduced.

The proof-of-work target expressed by CheckProofOfWork is modified to depend
on the timestamp offset t of the block itself.  This subtarget g starts at a
very low (difficult) value. It increases as the timestamp increases, and is
given by:

g = G(-1) * R * t^5

where:
g = Subtarget
G(-1) = Previous block's nominal target
R = 1 / 11036254599009
t = Timestamp - Timestamp(-1), capped at 2 days for recalculation

The calculation of the next nominal target G is also changed. G is given by:

G = G(-1) * t^6/6 * R / 531

Unlike traditional difficulty adjustment algorithms, this algorithm does not
depend on any historical blocktime offsets. It depends only on the target
immediately prior, and the blocktime offset of the block itself.

This algorithm is designed to maintain a mean inter-block time of 600 seconds,
even as hashrate and difficulty may vary dramatically. G(0) and minumum
difficulty are unchanged (0x1d00ffff for mainnet and testnet).

To retain the property that sibling blocks with differing timestamps are
equally preferred on the basis of chainwork, the block work comparator is
modified to use chainwork(-1) as the primary sort.  This is intuitive, because
timestamps now affect the nominal target one block sooner than before.